+++ backend/frontend-api/src/main/java/com/aprovaenem/gateway/FrontendApiApplication.java
@@ -2,11 +2,14 @@ package com.aprovaenem.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import reactor.core.publisher.Hooks;
 
 @SpringBootApplication
 public class FrontendApiApplication {
 
     public static void main(String[] args) {
+        Hooks.enableAutomaticContextPropagation();
         SpringApplication.run(FrontendApiApplication.class, args);
     }
 }
+